requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
python-dotenv==1.0.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
WTForms==3.1.2
gunicorn==22.0.0
